requirements.txt
bootstrap_flask==2.5.0
Flask==3.1.2
flask_sqlalchemy==3.1.1
flask_wtf==1.2.2
python-dotenv==1.2.1
SQLAlchemy==2.0.46
WTForms==3.2.1
